refactor(cli): dest decoding and content hashing leave import.go/status.go
Three helpers that every drift surface calls were reachable only by knowing
which command file they happened to land in. `decodeDestBytes`,
`jsonUnmarshalLoose` and `readDestFile` sat in `import.go` although `status`,
`diff`, `explain` and `reconcile` all read destinations through them;
`standardizeJSONC` — their JSONC dependency — sat in `status.go`, as did
`hashContent`/`hashFile`/`hashAnyValue` and the two `state.Key` constructors.

`destread.go` already exists and already documents itself as the one
destination-read gate for this package, so the decoders join the gate they
decode for. The hashes and their three opaque sentinels move to `hash.go`;
the two key constructors to `statekey.go`.

Pure relocation inside package cli: no identifier renamed, no body changed,
no import beyond each file's own block. `git diff --color-moved=zebra` shows
only the two new files' headers as non-moved lines.
